Refactor AMR and CPML examples, clean up imports and code structure
